+++ Changelog.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Add support for enums with associated values. DEV-17613.
+
+### Changed
+
+- Move the consumer type outputs into a trait for consistency/DRYing.
+
+## [0.6.0] - 2021-09-20
+
+### Changed
+
+- Update crate versions to fix dependency management in downstream crates.
+
+## [0.5.1] - 2021-09-20
+
+### Changed
+
+- Fix tests.
+
 ## [0.5.0] - 2021-09-22
 
 ### Added